functions/ingest/requirements.txt
crowdstrike-foundry-function==1.1.2
crowdstrike-falconpy==1.5.4
